Initial layout of the source code
Lua implementation of a lexer and a parser for the Lua language. The code
is an straightforward translation of the LuaJIT2's implementation.

The bytecode generator is equally implemented in Lua and is based on the
Nyanga's language implementation of Richard Hundt.

The parser call the AST builder when a grammar element is parsed. At the
end an AST tree is produced and passed to the bytecode generator. This
latter generate the bytecode corresponding to the given AST tree. The
bytecode can be run using LuaJIT with loadstring.