Implement correctly goto statements in bytecode generator
With the new implementation the UCLO instruction is generated instead of
JMP when required to close upvalues referenced by a closure.

The need_uclo flag to indicate that some variables needs UCLO
is now attached to labels to instead of scope. The better resolution is
needed to allow accurate UCLO instruction generation for goto statements.