Fix problem with TGETS/TSETS instructions
Github issue:

https://github.com/franko/luajit-lang-toolkit/issues/6

The string operand of TGETS is limited to 8 bit. The size of the string
index is now checked and if it does not fit 8 bit a TGETV is generated
instead.

The fix also potentially affect object's method call because of the object
indexing with the method name.

Initial fix kindly contributed by Daniel Kolesa.