Fix the error in rbase jump operand for logical expressions
To fix the problem a major change in the calling arguments was
required for methods expr_toreg and test_emit. In addition the
method expr_tonextreg is introduced to simplify the logic of
the code.

Essentially the methods expr_toreg and test_emit gets a new
argument "jreg" that indicates the next free register a
logical expression should jump to when a store is done.
The argument "jreg" will be equal to "freereg" if the current
expr_toreg is going to write of an register already allocated.
Otherwise, if expr_toreg is going to write to a new register
"jreg" will point to the next register.

Now the rbase argument for the JMP instructions is always
supposed to be correct and to be the same of those generated
by luajit.