Fix a bug due to missing FFI flag in bytecode dump
From lj_bcread.c from luajit it seems that whenever a KCDATA is used in
a function prototype or any of its child an FFI flag should be activated
in the generated bytecode.

Previously the "flags" word was always zero. Now the FFI flags is activated
whenever a KCDATA is used.