First incomplete implementation of bcread function
Implements in Lua a function to read a raw bytecode string and print
the bytes side by side with textual annotations.

The goal is to be able to easily inspect bytecode binary blobs byte by
byte with useful informations about the meaning of each byte.

The implementation is based on LuaJIT2 lj_bcread.c implementation file.

Currently only the bcread_header function is implemented.