Implement decoding of bytecode instructions
Based on "bcline" function from luajit's "jit.bc" module.

Current implementation does not reference the string, constants and
function operands, neither the upvalues and jump targets are not
highlited.

The current implementation is limited by the fact that the bytecode
section appears before the kgc and knum sections and it is therefore
impossible to resolve the constant references with a single pass.