Implement scoped lexical id generator
The new id generator creates identifier names based on a given name.

The name is initially prefixed with "@" to avoid conflicts with ordinary
variable names. Then, each time a scope is closed the name of generated
variable is changed removing the "@" and ensuring that there is no
conflicts with the variables in the scope.